fix: reorganizing services to keep things simple
