+++ plugins/dependency-packages-backend/src/services/DependencyTypeRegistry.ts
