+++ plugins/dependency-packages-backend/src/services/createDependencyTypeRegistration.ts
