--- plugins/dependency-packages-backend/src/services/dependencyType/index.ts
@@ -1,2 +1,0 @@
-export * from './createDependencyTypeRegistration';
-export * from './DependencyTypeRegistry';
