--- plugins/dependency-packages-backend/src/services/ownership/index.ts
@@ -1,1 +1,0 @@
-export { retrieveDefaultOwnerEntity } from './retrieveDefaultOwnerEntity';
